refactor: SkillMessage を components/ui/ に昇格
skill/rule 2ルートから共有されるため、CLAUDE.md 規約に従い
routes/skill/ → components/ui/ へ移動。
