+++ frontend/app/routes/rule/sections/OtherSection.tsx
@@ -1,6 +1,6 @@
 import { Link } from "react-router";
 
-import { SkillMessage } from "~/routes/skill/SkillMessage";
+import { SkillMessage } from "~/components/ui/SkillMessage";
 
 export function OtherSection() {
   return (
